Priority case: FLO brought from outside
FLO brought from outside the source will be processed after those brought from source-market.

+++ src/market.js
@@ -177,24 +177,20 @@ function initiateCoupling() {
 }
 
 function processCoupling(bestPairQueue) {
-    bestPairQueue.get().then(result => {
-        let buyer_best = result.buyOrder,
-            seller_best = result.sellOrder;
+    bestPairQueue.get().then(pair_result => {
+        let buyer_best = pair_result.buyOrder,
+            seller_best = pair_result.sellOrder;
         console.debug("Sell:", seller_best);
         console.debug("Buy:", buyer_best);
-        spendFLO(buyer_best, seller_best).then(txQueries => {
-            //process the Txn
-            var tx_quantity;
-            if (seller_best.quantity > buyer_best.quantity)
-                tx_quantity = processBuyOrder(seller_best, buyer_best, txQueries);
-            else if (seller_best.quantity < buyer_best.quantity)
-                tx_quantity = processSellOrder(seller_best, buyer_best, txQueries);
-            else
-                tx_quantity = processBuyAndSellOrder(seller_best, buyer_best, txQueries);
+        spendFLO(buyer_best, seller_best, pair_result.null_base).then(spend_result => {
+            let tx_quantity = spend_result.quantity,
+                txQueries = spend_result.txQueries,
+                clear_sell = spend_result.incomplete && !spend_result.flag_baseNull; //clear_sell can be true only if an order is placed without enough FLO
+            processOrders(seller_best, buyer_best, txQueries, tx_quantity, clear_sell);
             updateBalance(seller_best, buyer_best, txQueries, bestPairQueue.cur_rate, tx_quantity);
             //process txn query in SQL
-            DB.transaction(txQueries).then(results => {
-                bestPairQueue.next();
+            DB.transaction(txQueries).then(_ => {
+                bestPairQueue.next(quantity, spend_result.incomplete, spend_result.flag_baseNull);
                 console.log(`Transaction was successful! BuyOrder:${buyer_best.id}| SellOrder:${seller_best.id}`);
                 //Since a tx was successful, match again
                 processCoupling(bestPairQueue);
@@ -208,49 +204,46 @@ function processCoupling(bestPairQueue) {
     });
 }
 
-function spendFLO(buyOrder, sellOrder) {
+function spendFLO(buyOrder, sellOrder, null_base) {
     return new Promise((resolve, reject) => {
         DB.query("SELECT id, quantity, base FROM Vault WHERE floID=? ORDER BY base", [sellOrder.floID]).then(result => {
             let rem = Math.min(buyOrder.quantity, sellOrder.quantity),
-                txQueries = [];
-            for (let i = 0; i < result.length && rem > 0; i++) {
-                if (rem < result[i].quantity) {
-                    txQueries.push(["UPDATE Vault SET quantity=quantity-? WHERE id=?", [rem, result[i].id]]);
-                    rem = 0;
-                } else {
-                    txQueries.push(["DELETE FROM Vault WHERE id=?", [result[i].id]]);
-                    rem -= result[i].quantity;
-                }
-            }
-            if (rem > 0)
-                reject(`Sell order ${sellOrder.id} is active, but FLO is insufficient`);
-            else
-                resolve(txQueries);
+                txQueries = []
+            flag_baseNull = false;
+            for (let i = 0; i < result.length && rem > 0; i++)
+                if (result[i].base || null_base) {
+                    if (rem < result[i].quantity) {
+                        txQueries.push(["UPDATE Vault SET quantity=quantity-? WHERE id=?", [rem, result[i].id]]);
+                        rem = 0;
+                    } else {
+                        txQueries.push(["DELETE FROM Vault WHERE id=?", [result[i].id]]);
+                        rem -= result[i].quantity;
+                    }
+                } else
+                    flag_baseNull = true;
+            resolve({
+                quantity: Math.min(buyOrder.quantity, sellOrder.quantity) - rem,
+                txQueries,
+                incomplete: rem > 0,
+                flag_baseNull
+            });
         }).catch(error => reject(error));
     })
 }
 
-function processBuyOrder(seller_best, buyer_best, txQueries) {
-    let quantity = buyer_best.quantity;
-    //Buy order is completed, sell order is partially done.
-    txQueries.push(["DELETE FROM BuyOrder WHERE id=?", [buyer_best.id]]);
-    txQueries.push(["UPDATE SellOrder SET quantity=quantity-? WHERE id=?", [quantity, seller_best.id]]);
-    return quantity;
-}
-
-function processSellOrder(seller_best, buyer_best, txQueries) {
-    let quantity = seller_best.quantity;
-    //Sell order is completed, buy order is partially done.
-    txQueries.push(["DELETE FROM SellOrder WHERE id=?", [seller_best.id]]);
-    txQueries.push(["UPDATE BuyOrder SET quantity=quantity-? WHERE id=?", [quantity, buyer_best.id]]);
-    return quantity;
-}
-
-function processBuyAndSellOrder(seller_best, buyer_best, txQueries) {
-    //Both sell order and buy order is completed
-    txQueries.push(["DELETE FROM SellOrder WHERE id=?", [seller_best.id]]);
-    txQueries.push(["DELETE FROM BuyOrder WHERE id=?", [buyer_best.id]]);
-    return seller_best.quantity;
+function processOrders(seller_best, buyer_best, txQueries, quantity, clear_sell) {
+    if (quantity > buyer_best.quantity || quantity > seller_best.quantity)
+        throw Error("Tx quantity cannot be more than order quantity");
+    //Process Buy Order
+    if (quantity == buyer_best.quantity)
+        txQueries.push(["DELETE FROM BuyOrder WHERE id=?", [buyer_best.id]]);
+    else
+        txQueries.push(["UPDATE BuyOrder SET quantity=quantity-? WHERE id=?", [quantity, buyer_best.id]]);
+    //Process Sell Order
+    if (quantity == seller_best.quantity || clear_sell)
+        txQueries.push(["DELETE FROM SellOrder WHERE id=?", [seller_best.id]]);
+    else
+        txQueries.push(["UPDATE SellOrder SET quantity=quantity-? WHERE id=?", [quantity, seller_best.id]]);
 }
 
 function updateBalance(seller_best, buyer_best, txQueries, cur_price, quantity) {
@@ -262,7 +255,6 @@ function updateBalance(seller_best, buyer_best, txQueries, cur_price, quantity) 
     txQueries.push(["INSERT INTO Vault(floID, base, quantity) VALUES (?, ?, ?)", [buyer_best.floID, cur_price, quantity]])
     //Record transaction
     txQueries.push(["INSERT INTO Transactions (seller, buyer, quantity, unitValue) VALUES (?, ?, ?, ?)", [seller_best.floID, buyer_best.floID, quantity, cur_price]]);
-    return;
 }
 
 function getAccountDetails(floID) {
